+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Avoid playing the same songs over and over ([#35]).
 - Fix off-by-one check in currency transfer (`!currency give <user> <amount>`).
 - Only non-moderator and non-streamer chat bumps the idle counter.
+- URL whitelist is now stored in a setting `irc/whitelisted-hosts` ([#37]).
 
 ### Added
 - `!admin version` to check current setmod-bot package version. ([#32])